feat: move chart to Request Table block
