+++ my-app/src/components/RequestTable/RequestTable.tsx
@@ -7,14 +7,20 @@ import Table from "./Table/Table"
 import Footer from "./Footer/Footer";
 import EmptyContent from "./EmptyContent/EmptyContent";
 import SortingPanel from "./SortingPanel/SortingPanel";
+import Chart from "./Chart/Chart";
 
 function RequestTable() {
     const requests = useAppSelector((state) => state.requests.requests);
 
     const [tableData, setTableData] = useState(requests);
-    const [isVisible, setIsVisible] = useState(false);
+    const [isChartVisible, setIsChartVisible] = useState(false);
+    const [isTableVisible, setIsTableVisible] = useState(false);
 
     const {page, onChangePage, pagesCount, currentData} = usePagination(tableData, 15)
+
+    function switchChartVisibility() {
+        setIsChartVisible(prev => !prev)
+    }
 
     useEffect(() => {
         if (!isArrEmpty(requests)) {
@@ -23,24 +29,26 @@ function RequestTable() {
     }, [requests]);
 
     useEffect(() => {
-        !isArrEmpty(currentData) ? setIsVisible(true) : setIsVisible(false)
+        !isArrEmpty(currentData) ? setIsTableVisible(true) : setIsTableVisible(false)
     }, [currentData]);
 
     return (
         <Card w="65vw" h="90vh" shadow="sm" p="lg" radius="md" withBorder>
-            {isVisible
-                ? <>
-                    <SortingPanel data={requests} onChangeData={setTableData}/>
+            <SortingPanel data={requests} onChangeData={setTableData} isChartVisible={isChartVisible} switchChartVisibility={switchChartVisibility}/>
 
-                    <Table fullData={requests} currentData={currentData}/>
+            {isChartVisible
+                ? <Chart/>
+                : isTableVisible
+                    ? <>
+                        <Table fullData={requests} currentData={currentData}/>
 
-                    <Footer fullData={requests}
-                            currentPage={page}
-                            setCurrentPage={onChangePage}
-                            pagesCount={pagesCount}
-                    />
-                </>
-                : <EmptyContent content={'Nothing To Show'}/>
+                        <Footer fullData={requests}
+                                currentPage={page}
+                                setCurrentPage={onChangePage}
+                                pagesCount={pagesCount}
+                        />
+                    </>
+                    : <EmptyContent content={'Nothing to Show'}/>
             }
         </Card>
     );
