+++ my-app/src/components/RequestTable/SortingPanel/SortingPanel.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {Button, SimpleGrid} from "@mantine/core";
 import {SortingPanelProps, sortModeType} from "./SortingPanel.types";
 
-function SortingPanel({data, onChangeData}: SortingPanelProps) {
+function SortingPanel({data, onChangeData, isChartVisible, switchChartVisibility}: SortingPanelProps) {
     function sortByAmount(mode: sortModeType) {
         const arr = [...data];
 
@@ -17,17 +17,21 @@ function SortingPanel({data, onChangeData}: SortingPanelProps) {
     }
 
     return (
-        <SimpleGrid h="5%" mb="sm" spacing="xl" cols={3}>
-            <Button onClick={() => sortByAmount(1)} uppercase>
+        <SimpleGrid h="5%" mb="sm" spacing="xl" cols={4}>
+            <Button disabled={isChartVisible} onClick={() => sortByAmount(1)} uppercase>
                 Sort Amount by Increase
             </Button>
 
-            <Button onClick={() => sortByAmount(-1)} uppercase>
+            <Button disabled={isChartVisible} onClick={() => sortByAmount(-1)} uppercase>
                 Sort Amount by Decrease
             </Button>
 
-            <Button onClick={() => sortByAmount(0)} uppercase>
+            <Button disabled={isChartVisible} onClick={() => sortByAmount(0)} color="cyan" uppercase>
                 Default
+            </Button>
+
+            <Button onClick={switchChartVisibility} color="grape" uppercase>
+                {isChartVisible ? "Show Table" : "Show Chart"}
             </Button>
         </SimpleGrid>
     );
