+++ my-app/src/components/Tickers/Ticker/Form/Form.types.ts
@@ -1,4 +1,4 @@
-import {FormSubmitValues} from "../../../types/globalTypes";
+import {FormSubmitValues} from "../../../../types/globalTypes";
 
 export interface FormValues {
     instrument: string,
