+++ my-app/src/components/Tickers/Ticker/Ticker.types.ts
