+++ my-app/src/components/Tickers/Ticker/meta.ts
