+++ my-app/src/helpers/getInstrLabel.ts
@@ -1,4 +1,4 @@
-import {instruments} from "../components/Ticker/meta";
+import {instruments} from "../components/Tickers/Ticker/meta";
 
 export const getInstrLabel = (value: string): string => {
     const currentInst = instruments.filter(inst => inst.value === value)[0]
